% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoTRA.R
\docType{package}
\name{PoTRA}
\alias{PoTRA}
\alias{PoTRA-package}
\title{PoTRA Package}
\description{
The PoTRA analysis is based on topological ranks of genes in biological pathways. PoTRA can be used to detect pathways involved in disease. The PoTRA package contains one function for creating the PoTRA results object.
}
\details{
results <- PoTRA.corN(mydata=mydata,
                      genelist=genelist,
                      Num.sample.normal=Num.sample.normal,
                      Num.sample.case=Num.sample.case,
                      Pathway.database=Pathway.database,
                      PR.quantile=PR.quantile)

Pathway.database (options):
humanKEGG = pathways('hsapiens','kegg'),
humanBiocarta = pathways('hsapiens', 'biocarta'),
humanPharmGKB = pathways('hsapiens', 'pharmgkb').

mydata is a dataframe that contains rownames consisting of entrez gene identifiers and columns representing normals and samples from gene expression data.
The object genelist is a dataframe that consists of a single column of entrez gene identifiers (the same as those found in the rownames(mydata)).
Other objects such as Num.sample.normal and Num.sample.case represent normals and cases in the mydata dataframe. The Pathway.database object contains gene
lists from KEGG, Biocarta and PharmGKB databases, and these are made available through the graphite library.
The PR.quantile object contains the percentile of PageRank scores as a cutoff for hub genes, the recommended percentile is 0.95.
}
\author{
Chaoxing Li, Li Liu, Valentine Dinu \email{valentine.dinu@asu.edu}
}
