% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potra_man.R
\docType{package}
\name{PoTRA}
\alias{PoTRA.corN}
\alias{PoTRA-package}
\title{PoTRA.corN}
\description{The PoTRA analysis is based on topological ranks of genes in biological pathways. PoTRA can be used to detect pathways involved in disease. The PoTRA package contains one function for creating the PoTRA results object.}

\usage{PoTRA.corN(mydata, genelist, Num.sample.normal, Num.sample.case, Pathway.database, PR.quantile)}

\arguments{
  \item{mydata}{Dataframe that contains rownames consisting of entrez gene identifiers and columns representing normals and samples from gene expression data.}
  \item{genelist}{Dataframe that consists of a single column of entrez gene identifiers (the same as those found in the rownames(mydata)).}
  \item{Num.sample.normal}{Represents normal samples in the mydata dataframe.}
  \item{Num.sample.case}{Represents case samples in the mydata dataframe.}
  \item{Pathway.database}{Object contains gene lists from KEGG, Reactome, Biocarta and PharmGKB databases, and these are made available through the devel graphite library.}
  \item{PR.quantile}{Object contains the percentile of PageRank scores as a cutoff for hub genes, the recommended percentile is 0.95.}
}

\examples{ 

# Define the Pathway.database variable using the KEGG, Reactome, Biocarta and PharmGKB databases.
Pathway.database = humanKEGG

humanKEGG = pathways('hsapiens','kegg')
humanReactome = pathways('hsapiens','reactome')
humanBiocarta = pathways('hsapiens','biocarta')
humanPharmGKB = pathways('hsapiens','pharmgkb')
}  

\author{
Chaoxing Li, Li Liu, Valentine Dinu \email{valentine.dinu@asu.edu}
}
